Bootstrap from v0.7.1

+++ Makefile
@@ -1,6 +1,6 @@
 VERSION=${shell sh -c 'cat .version 2> /dev/null || git --git-dir git/fury/.git describe --exact-match --tags 2> /dev/null || git --git-dir git/fury/.git rev-parse --short HEAD'}
 BLOOPVERSION=1.3.5
-FURYSTABLE=0.6.7
+FURYSTABLE=0.7.1
 FURYLOCAL=opt/fury-$(FURYSTABLE)/bin/fury
 BINDEPS=coursier ng.py ng
 NAILGUNJAR=nailgun-server-1.0.0.jar
@@ -26,7 +26,7 @@ opt:
 	mkdir -p opt
 
 opt/fury-$(FURYSTABLE).tar.gz: opt
-	curl -s -o $@ "https://storage.googleapis.com/revivalist/downloads/fury.build/fury-$(FURYSTABLE).sh"
+	if [ ! -f "$@" ]; then curl -s -o $@ "https://storage.googleapis.com/revivalist/downloads/fury.build/fury-$(FURYSTABLE).sh"; fi
 
 $(FURYLOCAL): opt/fury-$(FURYSTABLE).tar.gz
 	sh opt/fury-$(FURYSTABLE).tar.gz opt/fury-$(FURYSTABLE)
@@ -64,6 +64,7 @@ dist/bundle/lib/$(NAILGUNJAR): dist/bundle/lib
 	curl -s -o $@ http://central.maven.org/maven2/com/facebook/nailgun-server/1.0.0/nailgun-server-1.0.0.jar
 
 dist/bundle/lib/fury-frontend.jar: dist/bundle/lib $(FURYLOCAL) bootstrap/bin .version src/**/*.scala
+	$(FURYLOCAL) standalone layer extract -f build.fury
 	$(FURYLOCAL) standalone permission grant --module frontend --project fury -P 729
 	$(FURYLOCAL) standalone build save --https --output linear --project fury --module frontend --dir $<
 	jar -uf $@ .version
